refactor: moved out full DMS Rules validation
